+++ releases.md
@@ -1,5 +1,9 @@
 # Releases
 
+## Unreleased
+
+  - **Breaking** Remove support for JavaScript packages installed in `lib/components`; use `node_modules` instead.
+
 ## v3.0.0
 
 The 3.0.x series is considered a development release while the protocol HTTP application and controller interfaces stabilize.